Enable asset precompilation without a database
This commit enables asset precompliation without access to a database.
This is useful when building deployable artefacts such as Docker images
with assets included to avoid a delay at launch to complete this step.

Set `ASSETS_PRECOMPILE` in the environment to skip various initialization
steps that require a database connection.

Co-authored-by: Graeme Porteous <graeme@rgbp.co.uk>